Codex/flow filter window modbus

+++ Core/Src/tim.c
@@ -22,7 +22,7 @@
 #include "key.h"
 #include "bsp_menu.h"
 /* USER CODE BEGIN 0 */
-volatile uint8_t  Timer3Uart1TimeBase10ms;
+volatile uint16_t Timer3Uart1TimeBase10ms;
 volatile uint8_t  DisplayTimeBase;
 volatile uint8_t  Timer3Uart2TimeBase10ms;
 volatile uint8_t  Timer3InitEnabled;
